Modularized the code and made anabstract class to avoid code duplication
